refactor(TextEditor): rename to RichTextEditor
