+++ packages/picasso-lab/src/RichTextEditor/hooks/useTextEditorState/test.ts
@@ -2,7 +2,7 @@ import { renderHook } from '@testing-library/react-hooks'
 import { act } from '@toptal/picasso/test-utils'
 
 import useTextEditorState from './useTextEditorState'
-import { initialState } from './../../store'
+import { initialState } from '../../store'
 import { actionTypes } from '../../store/toolbar'
 
 describe('useTextEditorState', () => {
