+++ packages/picasso-lab/src/RichTextEditor/store/toolbar/index.ts
