+++ packages/picasso-lab/src/RichTextEditor/store/toolbar/initialState.ts
