+++ packages/picasso-lab/src/RichTextEditor/types.ts
@@ -1,0 +1,3 @@
+import { ChangeHandler } from '../QuillEditor'
+
+export type RichTextEditorChangeHandler = ChangeHandler
