+++ packages/picasso-lab/src/RichTextEditorButton/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RichTextEditorButton'
