+++ packages/picasso-lab/src/RichTextEditorToolbar/index.ts
@@ -1,0 +1,2 @@
+export { default } from './RichTextEditorToolbar'
+export * from './types'
