--- packages/picasso-lab/src/TextEditor/types.ts
@@ -1,3 +1,0 @@
-import { ChangeHandler } from '../QuillEditor'
-
-export type TextEditorChangeHandler = ChangeHandler
