--- packages/picasso-lab/src/TextEditorButton/index.ts
@@ -1,1 +1,0 @@
-export { default } from './TextEditorButton'
