--- packages/picasso-lab/src/TextEditorToolbar/index.ts
@@ -1,2 +1,0 @@
-export { default } from './TextEditorToolbar'
-export * from './types'
