+++ packages/picasso-lab/src/index.ts
@@ -31,8 +31,8 @@ export { default as Section } from './Section'
 export type { SectionProps } from './Section'
 export { default as Note } from './Note'
 export type { NoteProps } from './Note'
-export { default as TextEditor } from './TextEditor'
-export type { TextEditorProps, TextEditorChangeHandler } from './TextEditor'
+export { default as RichTextEditor } from './RichTextEditor'
+export type { RichTextEditorProps, RichTextEditorChangeHandler } from './RichTextEditor'
 export { default as Timeline } from './Timeline'
 export type { TimelineProps } from './Timeline'
 // hygen code generator inserts export statements above this comment.
